More cleanup, revert the isPreview thing that is NOT fixed in Catalina, unlike what I thought. Should not have closed that radar :(
